Added Prerendering to the entities and added Restart Level
Prerendering applied to layers in Tiled that are heavy and may lead to
performance issues.  Prerendering them allows them to be loaded on an
offscreen canvas instead of being dynamically loaded as you go(default).
Pressing R restars the level tried to put in a fade but its weird so I
took it off